fixed push notification start, updated localization, added functionality with address manipulation
