+++ Multy/Extensions/Constants.swift
@@ -7,101 +7,6 @@ import UIKit
 private typealias LocalizeDelegate = Constants
 
 struct Constants {
-    static let yesString = "YES STRING"
-    static let noString = "NO STRING"
-    
-    static let cryptoPortfolioString = "CRYPTO PORTFOLIO"
-    static let currenciesChartsString = "CURRENCIES CHARTS"
-    static let jailbrokenDeviceWarningString = "JAILBROKEN DEVICE ALERT MESSAGE"
-    static let backupNeededString = "BACKUP IS NEEDED"
-    static let restoreMultyString = "RESTORE MULTY"
-    static let gettingWalletString = "GETTING WALLET"
-    static let creatingWalletString = "CREATING WALLET"
-    static let checkingVersionString = "CHECKING VERSION"
-    static let availableString = "AVAILABLE"
-    static let workInProgressString = "WORK IN PROGRESS"
-    static let weHaveUpdateString = "WE HAVE AN UPDATE"
-    static let updateMultyString = "UPDATE MULTY"
-    static let goToUpdateString = "GO TO UPDATE"
-    static let warningString = "WARNING"
-    static let cancelString = "CANCEL"
-    static let errorString = "ERROR"
-    static let walletNameAlertString = "WALLET NAME NOT EMPTY ALERT"
-    static let walletAmountAlertString = "WALLET AMOUNT ALERT"
-    static let deleteWalletAlertString = "DELETE WALLET WARNING"
-    static let copyToClipboardString = "COPY TO CLIPBOARD"
-    static let shareString = "SHARE"
-    static let magicalReceiveString = "MAGICAL RECEIVE"
-    static let transactionInfoString = "TRANSACTION INFO"
-    static let confirmationString = "CONFIRMATION"
-    static let confirmationsString = "CONFIRMATIONS"
-    static let notValidAddressString = "NOT VALID ADDRESS"
-    static let goToSettingsString = "GO TO SETTINGS"
-    static let recentAddressesString = "RECENT ADDRESSES"
-    static let veryFastString = "VERY FAST";
-    static let fastString = "FAST";
-    static let mediumString = "MEDIUM";
-    static let slowString = "SLOW";
-    static let verySlowString = "VERY SLOW";
-    static let customString = "CUSTOM";
-    static let predefinedValueMessageString = "PREDEFINED VALUE MESSAGE"
-    static let pleaseChooseFeeRate = "PLEASE CHOOSE FEE RATE"
-    static let moreThenYouHaveString = "MORE THEN YOU HAVE"
-    static let feeRateLessThenString = "FEE RATE LESS THEN"
-    static let satoshiPerByteString = "SATOSHI PER BYTE"
-    static let satoshiPerByteShortString = "SAT/B"
-    static let feeTooHighString = "FEE TOO HIGH"
-    static let gasPriceLess1String = "GAS PRICE LESS 1"
-    static let successString = "SUCCESS"
-    static let doneString = "DONE"
-    static let errorSendingTxString = "ERROR WHILE SENDING TX"
-    static let transactionErrorString = "TRANSACTION ERROR"
-    static let backupSeedPhraseNeededString = "BACKUP NEEDED"
-    static let viewSeedPhraseString = "VIEW SEED PHRASE"
-    static let areYouSureString = "ARE YOU SURE"
-    static let continueString = "CONTINUE"
-    static let nextWordString = "NEXT WORD"
-    static let from15String = "FROM 15"
-    static let orString = "OR"
-    static let wantToCancelString = "WANT TO CANCEL"
-    static let repeatPINCodeString = "REPEAT PIN-CODE"
-    static let sorryString = "SORRY"
-    static let noInternetString = "NO INTERNET"
-    static let connectDeviceString = "CONNECT DEVICE"
-    static let somethingWentWrongString = "SOMETHING WENT WRONG"
-    static let loadingString = "LOADING"
-    static let unableToSendString = "UNABLE SEND"
-    static let receiveString = "RECEIVE"
-    static let activeRequestsString = "ACTIVE REQUESTS"
-    static let youDontHaveWalletString = "YOU DONT HAVE WALLETS"
-    static let enterSatoshiPerByte = "ENTER SATOSHI PER BYTE"
-    static let enableBluetoothAlertTitle = "ENABLE BLUETOOTH ALERT TITLE"
-    static let settingsActionTitle = "SETTINGS ACTION TITLE"
-    static let sendTipString = "SEND TIP"
-    static let useTermsOfService = "USE TERMS OF SERVICE"
-    static let restoringWalletsString = "RESTORING WALLETS"
-    static let cannotChooseEmptyWalletString = "CANNOT CHOOSE EMPTY WALLET"
-    static let enterNonZeroAmountString = "ENTER NON-ZERO AMOUNT"
-    static let youCantEnterSumString = "YOU CANT ENTER SUM"
-    static let youEnteredTooSmallAmountString = "YOU ENTERED TOO SMALL AMOUNT"
-    static let youTryingSpendMoreThenHaveString = "YOU ARE TRYING SPEND MORE THEN HAVE"
-    static let updatingString = "UPDATING"
-    static let sendingString = "SENDING"
-    static let deletingString = "DELETING"
-    static let agreeWithTermsString = "AGREE WITH TERMS"
-    static let scanningNotSupportedString = "SCANNING NOT SUPPORTED"
-    static let deviceNotSupportingString = "DEVICE NOT SUPPORT"
-    static let notEnoughAmountString = "NOT ENOUGH AMOUNT"
-    static let pleaseChooseGasPriceString = "PLEASE CHOOSE GAS PRICE"
-    static let youCanUsePredefinedValueString = "YOU CAN USE PREDEFINED VALUE"
-    static let trySendZeroString = "TRY TO SEND ZERO"
-    static let enterCorrectValueString = "ENTER CORRECT VALUE"
-    static let noFundsString = "NO FUNDS"
-    static let errorWhileCreatingWalletString = "ERROR WHILE CREATING WALLET"
-    static let youCantSpendMoreThanFeeAndDonationString = "YOU CANT SPEND MORE THEN FEE AND DONATION"
-    static let youCantSpendMoreThanFeeString = "YOU CANT SPEND MORE THEN FEE"
-    static let andDonationString = "AND DONATION "
-    
     //Assets screen
     struct AssetsScreen {
         static let createWalletString = "Create wallet"
@@ -272,7 +177,7 @@ let minSatoshiToDonate: UInt64          = 5000  //5k minimum sum to donate
 //API REST constants
 //let apiUrl = "http://88.198.47.112:2278/"//"http://192.168.0.121:7778/"
 
-let shortURL = "api.multy.io"
+let shortURL = "test.multy.io"
 let apiUrl = "http://\(shortURL)/"
 let socketUrl = "ws://\(shortURL)/"
 //JACK
