+++ Multy/Storyboards/en.lproj/Wallet.strings
@@ -168,3 +168,6 @@
 "zRh-am-gEl.text" = "Exchange"
 
 "f6g-fm-Ou2.text" = "You don’t have any transactions yet. Create your first operation"
+
+"QeZ-MK-mx7.normalTitle" = "Add to Contact"
+
