+++ Multy/Storyboards/ru.lproj/Wallet.strings
@@ -168,3 +168,5 @@
 "zRh-am-gEl.text" = "Биржа"
 
 "f6g-fm-Ou2.text" = "Нажмите кнопку, чтобы создать запрос на получение средств."
+
+"QeZ-MK-mx7.normalTitle" = "Добавить в контакт"
