+++ Multy/Storyboards/uk.lproj/Wallet.strings
@@ -168,3 +168,5 @@
 "zRh-am-gEl.text" = "Біржа"
 
 "f6g-fm-Ou2.text" = "Натисніть кнопку, щоб створити запит на отримання коштів."
+
+"QeZ-MK-mx7.normalTitle" = "Додати в контакт"
