+++ Multy/Strings/en.lproj/Sends.strings
@@ -55,3 +55,4 @@
 "YOU CANT SPEND MORE THEN FEE AND DONATION" = "You can`t spend sum more than you have!\nDon`t forget about Fee and donation.\n\nYour fee is ";
 "YOU CANT SPEND MORE THEN FEE" = "You can`t spend sum more than you have!\nDon`t forget about Fee.\nYour fee is ";
 "AND DONATION " = "\nand donation is ";
+"ADD TO CONTACTS" = "Add to Contacts";
