+++ Multy/Strings/ru.lproj/Sends.strings
@@ -55,3 +55,4 @@
 "YOU CANT SPEND MORE THEN FEE AND DONATION" = "Вы не можете потратить больше, чем имеется кошельке! Не забывайте о комиссии и пожертвовании. Ваша комиссия - ";
 "YOU CANT SPEND MORE THEN FEE" = "Вы не можете потратить больше, чем имеется кошельке! Не забывайте о комиссии. Ваша комиссия - ";
 "AND DONATION " = ",\nпожертвование - ";
+"ADD TO CONTACTS" = "Добавить в контакт";
