+++ Multy/Strings/uk.lproj/Sends.strings
@@ -55,3 +55,4 @@
 "YOU CANT SPEND MORE THEN FEE AND DONATION" = "Ви не можете витратити більше, ніж є гаманці!\nНе забувайте про комісію та пожертвування. Ваша комісія - ";
 "YOU CANT SPEND MORE THEN FEE" = "Ви не можете витратити більше, ніж є гаманці! Не забывайте о комиссии. Ваша комісія - ";
 "AND DONATION " = ",\nпожертвування - ";
+"ADD TO CONTACTS" = "Додати в контакт";
