+++ Multy/ViewControllers and Presenters/Seed Phrase/Check Phrase/CheckWordsPresenter.swift
@@ -60,6 +60,7 @@ class CheckWordsPresenter: NSObject {
             self.checkWordsVC?.loader.hide()
             
             DataManager.shared.socketManager.start()
+            DataManager.shared.subscribeToFirebaseMessaging()
         }
     }
 }
